tasks/cephfs: improve fuse force umount
This is a separate thing from strengthening
the umount -f behaviour -- in this case, we
are specifically intending to force-unmount,
so we can kill the fuse process and hopefully
never hit the force-force-umount fallback.

Avoiding the fallback is a good thing because
it kills all the fuse processes on the host,
not just the one we are interested in.

Signed-off-by: John Spray <john.spray@redhat.com>